backend/AI/services/hs_service/requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
numpy
faiss-cpu
fastparquet
pyarrow
python-dotenv
